# Contacts-api/public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts-api",
    "license": {
      "name": "none"
    },
    "description": "This is a sample Node Rest API."
  },
  "host": "127.0.0.1:1337",
  "basePath": "/api/v1.0",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "text/plain",
    "application/json",
    "text/xml"
  ],
  "paths": {
...
    "/contacts/{contactId}/addresses/{addressId}": {
      "get": {
        "summary": "Select an address in a contact",
        "tags": [
          "contacts"
        ],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "Numeric ID of contact"
          },
          {
            "name": "addressId",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "Numeric ID of contact"
          }
        ],
        "responses": {
          "200": {
            "description": "Returned address",
            "headers": {
              "Connection": { "type": "string" },
              "Content-Length": { "type": "string" },
              "Content-Security-Policy": { "type": "string" },
              "Content-Type": { "type": "string" },
              "Date": { "type": "string" },
              "ETag": { "type": "string" },
              "Referrer-Policy": { "type": "string" },
              "Strict-Transport-Security": { "type": "string" },
              "Vary": { "type": "string" },
              "X-Content-Security-Policy": { "type": "string" },
              "X-Content-Content-Type-Options": { "type": "string" },
              "X-DNS-Prefetch-Control": { "type": "string" },
              "X-Download-Options": { "type": "string" },
              "X-Frame-Options": { "type": "string" },
              "X-Webkit-CSP": { "type": "string" },
              "X-XSS-Protection": { "type": "string" }
            },
            "schema": {
              "$ref": "#/definitions/Address"
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {
              "Connection": { "type": "string" },
              "Content-Length": { "type": "string" },
              "Content-Security-Policy": { "type": "string" },
              "Content-Type": { "type": "string" },
              "Date": { "type": "string" },
              "ETag": { "type": "string" },
              "Referrer-Policy": { "type": "string" },
              "Strict-Transport-Security": { "type": "string" },
              "X-Content-Security-Policy": { "type": "string" },
              "X-Content-Content-Type-Options": { "type": "string" },
              "X-DNS-Prefetch-Control": { "type": "string" },
              "X-Download-Options": { "type": "string" },
              "X-Frame-Options": { "type": "string" },
              "X-Webkit-CSP": { "type": "string" },
              "X-XSS-Protection": { "type": "string" }
            }
          },
          "5XX": {
            "description": "SERVER ERROR",
            "headers": {
              "Connection": { "type": "string" },
              "Content-Length": { "type": "string" },
              "Content-Security-Policy": { "type": "string" },
              "Date": { "type": "string" },
              "Referrer-Policy": { "type": "string" },
              "Strict-Transport-Security": { "type": "string" },
              "X-Content-Security-Policy": { "type": "string" },
              "X-Content-Content-Type-Options": { "type": "string" },
              "X-DNS-Prefetch-Control": { "type": "string" },
              "X-Download-Options": { "type": "string" },
              "X-Frame-Options": { "type": "string" },
              "X-Webkit-CSP": { "type": "string" },
              "X-XSS-Protection": { "type": "string" }
            }
          }
        }
      },
      "put": {
        "summary": "Update an address in a contact",
        "tags": [
          "contacts"
        ],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "Numeric ID of contact"
          },
          {
            "name": "addressId",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "Numeric ID of contact"
          },
          {
            "name": "Address",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Address" }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned address",
            "headers": {
              "Connection": { "type": "string" },
              "Content-Length": { "type": "string" },
              "Content-Security-Policy": { "type": "string" },
              "Content-Type": { "type": "string" },
              "Date": { "type": "string" },
              "ETag": { "type": "string" },
              "Referrer-Policy": { "type": "string" },
              "Strict-Transport-Security": { "type": "string" },
              "Vary": { "type": "string" },
              "X-Content-Security-Policy": { "type": "string" },
              "X-Content-Content-Type-Options": { "type": "string" },
              "X-DNS-Prefetch-Control": { "type": "string" },
              "X-Download-Options": { "type": "string" },
              "X-Frame-Options": { "type": "string" },
              "X-Webkit-CSP": { "type": "string" },
              "X-XSS-Protection": { "type": "string" }
            },
            "schema": {
              "$ref": "#/definitions/Address"
            }
          },
          "400": {
            "description": "Invalid input",
            "headers": {
              "Connection": { "type": "string" },
              "Content-Length": { "type": "string" },
              "Content-Security-Policy": { "type": "string" },
              "Content-Type": { "type": "string" },
              "Date": { "type": "string" },
              "ETag": { "type": "string" },
              "Referrer-Policy": { "type": "string" },
              "Strict-Transport-Security": { "type": "string" },
              "X-Content-Security-Policy": { "type": "string" },
              "X-Content-Content-Type-Options": { "type": "string" },
              "X-DNS-Prefetch-Control": { "type": "string" },
              "X-Download-Options": { "type": "string" },
              "X-Frame-Options": { "type": "string" },
              "X-Webkit-CSP": { "type": "string" },
              "X-XSS-Protection": { "type": "string" }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {
              "Connection": { "type": "string" },
              "Content-Length": { "type": "string" },
              "Content-Security-Policy": { "type": "string" },
              "Content-Type": { "type": "string" },
              "Date": { "type": "string" },
              "ETag": { "type": "string" },
              "Referrer-Policy": { "type": "string" },
              "Strict-Transport-Security": { "type": "string" },
              "X-Content-Security-Policy": { "type": "string" },
              "X-Content-Content-Type-Options": { "type": "string" },
              "X-DNS-Prefetch-Control": { "type": "string" },
              "X-Download-Options": { "type": "string" },
              "X-Frame-Options": { "type": "string" },
              "X-Webkit-CSP": { "type": "string" },
              "X-XSS-Protection": { "type": "string" }
            }
          },
          "5XX": {
            "description": "SERVER ERROR",
            "headers": {
              "Connection": { "type": "string" },
              "Content-Length": { "type": "string" },
              "Content-Security-Policy": { "type": "string" },
              "Date": { "type": "string" },
              "Referrer-Policy": { "type": "string" },
              "Strict-Transport-Security": { "type": "string" },
              "X-Content-Security-Policy": { "type": "string" },
              "X-Content-Content-Type-Options": { "type": "string" },
              "X-DNS-Prefetch-Control": { "type": "string" },
              "X-Download-Options": { "type": "string" },
              "X-Frame-Options": { "type": "string" },
              "X-Webkit-CSP": { "type": "string" },
              "X-XSS-Protection": { "type": "string" }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an address in a contact",
        "tags": [
          "contacts"
        ],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "Numeric ID of contact"
          },
          {
            "name": "addressId",
            "in": "path",
            "type": "integer",
            "required": true,
            "description": "Numeric ID of contact"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted address",
            "headers": {
              "Connection": { "type": "string" },
              "Content-Security-Policy": { "type": "string" },
              "Date": { "type": "string" },
              "ETag": { "type": "string" },
              "Referrer-Policy": { "type": "string" },
              "Strict-Transport-Security": { "type": "string" },
              "X-Content-Security-Policy": { "type": "string" },
              "X-Content-Content-Type-Options": { "type": "string" },
              "X-DNS-Prefetch-Control": { "type": "string" },
              "X-Download-Options": { "type": "string" },
              "X-Frame-Options": { "type": "string" },
              "X-Webkit-CSP": { "type": "string" },
              "X-XSS-Protection": { "type": "string" }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {
              "Connection": { "type": "string" },
              "Content-Length": { "type": "string" },
              "Content-Security-Policy": { "type": "string" },
              "Content-Type": { "type": "string" },
              "Date": { "type": "string" },
              "ETag": { "type": "string" },
              "Referrer-Policy": { "type": "string" },
              "Strict-Transport-Security": { "type": "string" },
              "X-Content-Security-Policy": { "type": "string" },
              "X-Content-Content-Type-Options": { "type": "string" },
              "X-DNS-Prefetch-Control": { "type": "string" },
              "X-Download-Options": { "type": "string" },
              "X-Frame-Options": { "type": "string" },
              "X-Webkit-CSP": { "type": "string" },
              "X-XSS-Protection": { "type": "string" }
            }
          },
          "5XX": {
            "description": "SERVER ERROR",
            "headers": {
              "Connection": { "type": "string" },
              "Content-Length": { "type": "string" },
              "Content-Security-Policy": { "type": "string" },
              "Date": { "type": "string" },
              "Referrer-Policy": { "type": "string" },
              "Strict-Transport-Security": { "type": "string" },
              "X-Content-Security-Policy": { "type": "string" },
              "X-Content-Content-Type-Options": { "type": "string" },
              "X-DNS-Prefetch-Control": { "type": "string" },
              "X-Download-Options": { "type": "string" },
              "X-Frame-Options": { "type": "string" },
              "X-Webkit-CSP": { "type": "string" },
              "X-XSS-Protection": { "type": "string" }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Address": {
      "required": [
        "primary"
      ],
      "properties": {
        "primary": {
          "type": "boolean",
          "example": true
        },
        "addressL1": {
          "type": "string",
          "example": "Apt 23"
        },
        "addressL2": {
          "type": "string",
          "example": "Not The Sewers Street"
        },
        "city": {
          "type": "string",
          "example": "New New York"
        },
        "state": {
          "type": "string",
          "example": "NY"
        },
        "country": {
          "type": "string",
          "example": "The Government of Earth"
        },
        "zipcode": {
          "type": "string",
          "example": "12343-45"
        }
      }
    },
...
  }
}